Remove v1 encoding support for WIT documents/worlds
This commit removes support for the "v1" encoding of WIT packages into
WebAssembly. This was originally migrated to v2 in #1252 and v2 was
enabled by default in #1274. This finishes the transition by removing
support for creating the old encoding.
